// src/app/shared/models/insurance-place-information.ts
export interface InsurancePlaceInformation {
  zipCode: string;
  municipality: string; // Obec
  municipalityPart: string; // Cast obce
  street: string; // Ulice
  numberDescriptive: string; // Cislo popisne
  numberRegistered: string; // Cislo evidencni
  numberOriented: string; // Cislo orientacni
}